+++ d60.Cirqus.AzureServiceBus/Relay/AzureServiceBusRelayEventStoreProxy.cs
@@ -69,7 +69,7 @@ namespace d60.Cirqus.AzureServiceBus.Relay
         {
         }
 
-        public IEnumerable<Event> Load(string aggregateRootId, long firstSeq = 0)
+        public IEnumerable<EventData> Load(string aggregateRootId, long firstSeq = 0)
         {
             var transportMessage = InnerLoad(aggregateRootId, firstSeq);