+++ d60.Cirqus.PostgreSql/PostgreSqlEventStore.cs
@@ -168,7 +168,7 @@ INSERT INTO ""{0}"" (
             return connection;
         }
 
-        public IEnumerable<Event> Load(string aggregateRootId, long firstSeq = 0)
+        public IEnumerable<EventData> Load(string aggregateRootId, long firstSeq = 0)
         {
             using (var connection = GetConnection())
             {