+++ d60.Cirqus/MsSql/Events/MsSqlEventStore.cs
@@ -107,7 +107,7 @@ INSERT INTO [{0}] (
             }
         }
 
-        public IEnumerable<Event> Load(string aggregateRootId, long firstSeq = 0)
+        public IEnumerable<EventData> Load(string aggregateRootId, long firstSeq = 0)
         {
             SqlConnection conn = null;